% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/phylogeneticClades.R
\name{clades}
\alias{clades}
\title{Phylogenetic clades}
\usage{
clades(phylo, community, n)
}
\arguments{
\item{phylo}{An object of class 'phylo'}

\item{community}{A community matrix}

\item{n}{The number of clades to visualise}
}
\description{
Visualises the phylogenetic clades within a metacommunity from local community plots.
}

