% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rFunctions.R
\name{collessImbalance}
\alias{collessImbalance}
\title{Colless' imbalance}
\usage{
collessImbalance(phylo)
}
\arguments{
\item{phylo}{An object of class 'phylo'}
}
\value{
A numeric value for the Colless' Imbalance
}
\description{
Calculates the Colless' imbalance for a Phylogeny.
}

