% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createPhylosimList.R
\name{createPhylosimList}
\alias{createPhylosimList}
\title{PhylosimList creator}
\usage{
createPhylosimList(simu)
}
\arguments{
\item{simu}{Multiple objects of type Phylosim as created by \code{\link{runSimulation}}}
}
\value{
An object of type PhylosimList
}
\description{
Creates an object of class "PhylosimList" from a list of "Phylosim" objects.
}
\details{
The function is used to make functions for objets of type "PhylosimList" usable for objects of class "Phylosim". The function needs at least two objets to work.
}
\examples{
## Create two objects of type Phylosim
par1 <- createCompletePar(x = 50, y = 50, dispersal = "global" , runs = 500,
        density = 1, environment = 0.5, specRate = 1)
par2 <- createCompletePar(x = 50, y = 50, dispersal = "global" , runs = 500,
        density = 1, environment = 0.5, specRate = 2)
        
simu1 <- runSimulation(par1)
simu2 <- runSimulation(par2)

## Use the function
simuList <- createPhylosimList(list(simu1,simu2))

## The object can now be handed to functions designed for objects of type "PhylosimList".
 
}

