% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rFunctions.R
\name{es}
\alias{es}
\title{Expected species richness}
\usage{
es(specRate, dimensions)
}
\arguments{
\item{specRate}{The speciation rate in speciations per generation}

\item{dimensions}{The length of the grid edge}
}
\value{
A float value for the expected species richness
}
\description{
Calculate the expected species richness in the equilibrium state as calculated by Hubbel
}
\examples{
es(2, 100)
}

