% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/modelParallel.R
\name{fullModParallel}
\alias{fullModParallel}
\title{Species community Simulation}
\usage{
fullModParallel(parameters = NULL, cores = NULL, x = NULL, y = NULL,
  dispersal = NULL, runs = NULL, specRate = NULL, density = NULL,
  environment = NULL, mortalityFitness = NULL, mortalityStrength = NULL,
  reproductiveFitness = NULL, neutral = NULL, dispersalCut = NULL,
  densityCut = NULL, seed = NULL, saveLocation = FALSE, backup = FALSE)
}
\arguments{
\item{parameters}{.xml file containing all parameters for the model scenarios (default is NULL)}

\item{cores}{The number of cores to be used for parallel computing (default is NULL)}

\item{x}{Dimension of the model landscape in x-direction}

\item{y}{Dimension of the model landscape in y-direction}

\item{dispersal}{1 = Gloabl dispersal, 3 = Local dispersal}

\item{runs}{Number of generations the model runs over}

\item{density}{Logical determining whether or not density dependence influences the model}

\item{environment}{Logical determining whether or not the environment influences the model}

\item{mortalityFitness}{Logical, determining whether or not the individual fitness influences mortality}

\item{mortalityStrength}{Integer determining the influence of fitness on mortality}

\item{reproductiveFitness}{Logical, determining whether or not the individual fitness influences reproduction}

\item{neutral}{Logical determining whether or not the model is run under neutral conditions (overrides density and environment)}

\item{dispersalCut}{Integer defining the maximum dispersal distance of the kernel}

\item{densityCut}{Integer defining the range for the density dependence}

\item{seed}{Integer setting the random seed for the model}

\item{saveLocation}{Path of the folder to save the extracted files to}

\item{backup}{Logical determining whether the results of the individual scenario runs should be saved as a workspace image (advised if the simulation takes a long time, or if the individual scenarios vary greatly in runtime. Default is FALSE)}
}
\value{
A list containing numerical matrices of species distribution, local trait values and the environment for each scenario (each cell in a matrix represents an individual of a species), plus a phylogeny object of class "phylo" for each scenario.
}
\description{
A model of species community assembly under different assembly mechanisms, using parallel computing to make use of multi core cpus and clusters in order to reduce computation time.
}
\details{
This function uses the \code{\link{foreach}} and \code{\link{doParallel}} package to compute the model scenarios parallel on several cores. if you want to keep working on your computer make sure to reserve at least one core for your other endevors (by assigning n-1 cores to the function). By default all cores are employed to ensure maximum speed.\cr \cr The phylogeny is passed to R in the newick format and parsed to an object of class "phylo" with the function \code{\link[ape]{read.tree}} from the \code{\link{ape}} package. \cr \cr If no .xml parameter file is supplied, a single scenario can be computed by manually setting the parameters. \cr \cr If mortalityFitness = TRUE, the number of reproduction events per generation is doubled. The mortality strength is an Integer value which determines the number of events per generation in which the mortality depends on fitness. \cr E.g. if mortalityStrengt = 100 for every 100 events with fitness dependent mortality there is 1 run w
}

