% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/parametersXML.R
\name{getParametersXML}
\alias{getParametersXML}
\title{Get model parameters}
\usage{
getParametersXML(file)
}
\arguments{
\item{file}{The .xml file to be inspected}
}
\value{
A dataframe with the model parameters
}
\description{
Displays the model parameters from an .xml file as set by \code{\link{setModelParametersXML}} or \code{\link{setLeipzigParametersXML}}
}
\examples{
# Set the model parameters
#path <- "C:/test.xml"
#setParametersXML(x = 50, y = 50, runs = 3000, dispersal = 1, nSpec = 1, specRate = 2, density = T, densityStrength = 0.4, location = path)

# Retrive the model parameters
#modelParameters <- getParametersXML(path)
#modelParameters
}

