% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rWrappers.R
\name{leip}
\alias{leip}
\title{Species community model after chave et al (2000), as designed in leipzig}
\usage{
leip(params)
}
\arguments{
\item{x}{Dimension of the model landscape in x-direction}

\item{y}{Dimension of the model landscape in y-direction}

\item{dispersal}{1 = Gloabl dispersal, 2 = Nearest neighbor dispersal}

\item{nSpec}{Number of initial Species in the model landscape}

\item{specRate}{Numerical the number of species introduced into the community during each generation}

\item{runs}{Number of generations the model runs over}

\item{density}{Logical determining whether or not density dependence influences the model}

\item{densityStrength}{Numerical (0-1) determining the strength of the density dependence}
}
\value{
a numerical matrix (each cell represents an individual of a species)
}
\description{
A simple model of species community assembly following the design used by Chave et al. (2000).
}

