% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/leipzigParallel.R
\name{leipScenarios}
\alias{leipScenarios}
\title{Species community model scenarios (Leipzig version)}
\usage{
leipScenarios(params, cores = NULL)
}
\arguments{
\item{params}{Character, The .xml file containing the model parameters as set by \code{\link{setParametersXML}}}

\item{cores}{Integer, The number of cores reserved for parallel computing}
}
\value{
a list of numerical matrices of species distribution and local density values for each scenario (each cell in a matrix represents an individual of a species)
}
\description{
A function to calculate several scenarios of a simple model of species community assembly following the design used by Chave et al. (2000), using parallel computing.
}
\details{
This function uses the \pkg{foreach} and \pkg{doParallel} package to compute the model scenarios parallel on several cores. if you want to keep working on your computer make sure to reserve at least one core for your other endevors (by assigning n-1 cores to the function). By default all cores are employed to ensure maximum speed.
}
\examples{
# Define parameter values
scenarios = c("globalDens", "globalNoDens",
              "localDens", "localNoDens")
x = rep(50,4)
y = rep(50,4)
runs = rep(3000,4)
dispersal = c(1,1,2,2)
nSpec = rep(1,4)
specRate = rep(2,4)
density = c(T,F,T,F)
densityStrength =  rep(0.4,4)
path <- "C:/test.xml"

# Set parameters for the model
setLeipzigParametersXML(scenarios = scenarios, x = x, y = y, runs = runs,
                dispersal = dispersal, nSpec = nSpec,
                specRate = specRate,
                density = density, densityStrength = densityStrength,
                location = path)

# Run the model with the set parameters on 2 cores
Out <- leipScenarios(params = path, cores = 2)
}

