% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rFunctions.R
\name{localPlots}
\alias{localPlots}
\title{Metacommunity subplots}
\usage{
localPlots(size, n, matrix, community = F)
}
\arguments{
\item{size}{A single value or a vector determining the size(s) of the generated subplots}

\item{n}{The number of subplots to be generated}

\item{matrix}{A square matrix containing the metacommunity (one individual per grid cell)}

\item{community}{Logical, determining whther to generate a community matrix or not default is FALSE}
}
\value{
A list of subplots and (if coomunity = T) a community matrix with plots in rows and species in collumns
}
\description{
Generates subplots from a metacommunity, simulating field conditions
}

