% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dllExportTest.R
\name{multiInt}
\alias{multiInt}
\title{Integer multiplication}
\usage{
multiInt(x)
}
\arguments{
\item{x}{Integer value to be doubled}
}
\value{
An Integer.
}
\description{
multiplies an integer with 2
}

