% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/neutralModelR.R
\name{NeutralMod}
\alias{NeutralMod}
\title{Model of neutral community assembly}
\usage{
NeutralMod(xdim = 100, ydim = 100, specRate = 2, seed = NULL,
  runs = 500)
}
\arguments{
\item{specRate}{The speciation rate in total speciation events per generation}

\item{seed}{The random seed (do not change this option if you previously set the seed!)}

\item{runs}{The number of generations the model runs through}

\item{dim}{The edge-length of the square metacommunity matrix}
}
\description{
Runs a simple model of species community assembly under neutral conditions as defined by Hubbel.
}
\details{
Be careful with the dimensions you choose. Large grids and many generations may take very long to compute. You can assume to have reached the equilibrium state of the model after dim^2/2.
}
\examples{
# Run the model
metaCom <- NeutralMod(dim=50, specRate=2, seed=1500, runs=1500)
# Visualize the grid
image(metaCom)
# Calculate the species area-relation
rac(metaCom, plot=T)
}

