% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nullModel.R
\name{nullModel}
\alias{nullModel}
\title{Null Model}
\usage{
nullModel(speciesMatrix, phylogeny, localPlotSize, numberOfPlots, repetitions,
  fun = mpd)
}
\arguments{
\item{speciesMatrix}{A square matrix containing the metacommunity (one individual per grid cell)}

\item{phylogeny}{an object of class "phylo", representing the metacommunity}

\item{localPlotSize}{number of grid cells of the plots drawn within the metacommunity}

\item{numberOfPlots}{number of plots drawn within the metacommunity}

\item{repetitions}{number of generated null model plots}
}
\value{
A numeric vector with pValues for each plot in the observed metacommunity
}
\description{
Generates a null model to compare the observed results against.
}

