% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nullModelSample.R
\name{nullModelSample}
\alias{nullModelSample}
\title{Null Model (sample abundance)}
\usage{
nullModelSample(speciesMatrix, phylogeny, localPlotSize, numberOfPlots,
  repetitions)
}
\arguments{
\item{speciesMatrix}{A square matrix containing the metacommunity (one individual per grid cell)}

\item{phylogeny}{an object of class "phylo", representing the metacommunity}

\item{localPlotSize}{number of grid cells of the plots drawn within the metacommunity}

\item{numberOfPlots}{number of plots drawn within the metacommunity}

\item{repetitions}{number of generated null model plots}
}
\value{
A numeric vector with pValues for each plot in the observed metacommunity
}
\description{
Generates a null model based on the abundance distribution of the sample plots to compare the observed results against.
}

