\name{PhylGeo-package}
\alias{PhylGeo-package}
\alias{PhylGeo}
\docType{package}
\title{
Analysing phylogenetic and biogeographic patterns and inferring species community assembly 
}
\description{
This package is designed to simulate and analyse species community assembly processes. It allows the user to implement different mechanisms of community assembly such as environmental filtering, competitive exclusion or dispersal limitation. These mechanisms can be combined or activated seprately, which allows the user to identify biogeographic or phylogenetic patterns exclusively correlated with a certain mechanism or combination of mechanisms.
}
\details{
\tabular{ll}{
Package: \tab PhylGeo\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-02-12\cr
License: \tab GPL (>= 2)\cr
}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
Paul Bauche

Maintainer: Paul Bauche <your@email.com>
}
\references{
~~ Literature or other references for background information ~~
}

\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%% ~~ simple examples of the most important functions ~~
}
