% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/reconstructPhylogeny.R
\name{phyloReconstruct}
\alias{phyloReconstruct}
\title{Phxlogeny Reconstruction}
\usage{
phyloReconstruct(speciesMatrix, traitMatrix, ...)
}
\arguments{
\item{speciesMatrix}{A square matrix containing a species community (one individual per grid cell)}

\item{traitMatrix}{A square matrix containing the neutral trait of a species community (one trait value per individual per grid cell}

\item{...}{Additional arguments to be passed to function \code{\link{hclust}}}
}
\value{
An object of class 'phylo'
}
\description{
reconstructs a phylogeny from a neutral trait matrix as returned by \code{\link{fullMod}}
}

