% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotclades.R
\name{plotclades}
\alias{plotclades}
\title{Phylogenetic clades}
\usage{
plotclades(simu, which.simulation = NULL, size, n, subplot = NULL,
  phylogeny = NULL)
}
\arguments{
\item{simu}{Simulation output of the class "Phylosim", usually consisting out of several lists. Needs to contain at least the phylogeny ($phylogeny) and species matrix ($specMat)}

\item{which.simulation}{defines which simulation run to choose in case you defined to save at multiple time steps. The default is the last one.}

\item{n}{The number of local communities. For each community a new plot is created}

\item{subplot}{if no simu object is given, a communitytable of a local community can be given here}

\item{phylogeny}{if no simu object is ginve, an object of the class 'phylo' can be given here to describe the communites phylogeny}
}
\description{
Visualises the phylogenetic clades of a local community within a metacommunity.
}
\details{
a
}

