% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotPhylogeneticDispersion_newTestVersion.R
\name{plotPD}
\alias{plotPD}
\title{Plot Phylogenetic Dispersion}
\usage{
plotPD(pvalues, positions = NULL, title = "P-values")
}
\arguments{
\item{pvalues}{List of p-values as generated by \code{\link{calculatePhylogeneticDispersion}}.}

\item{positions}{List, determines where the plots should be placed and how they should be labeled (see Details). The list need to have the entries x, y, xname and yname.}

\item{title}{String, determines the title of the plot}
}
\description{
Plotting function for the results of \code{\link{calculatePhylogeneticDispersion}}. Usefull for plotting different scenarios.
}
\details{
To give meningful results, the function needs the results of a full factorial experimental design realized by \code{\link{calculatePhylogeneticDispersion}}. For an further explanation please see the Examples section.
}
\examples{
## First the experimental design has to be defined.
## As stated above you need to implement a full factorial design.
## The design here will lead to 28 experiments.
dispOptions = 4
fitOptions = 7

# These are the parameters that we want to test.
dispersal =  rep(c(0,0.5, 2,4), each = fitOptions)
density = rep(c(0,seq(0,1,len = fitOptions-2),1),dispOptions)

# Further an environmental effect is added.
environment = rep(c(0,seq(1,0,len = fitOptions-2),1),dispOptions)

## The modes and scenarios are only defined to label the plots. They have no influence on the
## calculations.
modes <- ifelse(dispersal == 0, "global", "local")
scenarios <- paste(modes, " dens=", density, " env=", environment, sep = "")

## Now we can define a list of parameter sets.
pars <- list()
for (i in 1:length(scenarios)){pars[[i]] = createCompletePar(x = 20,y = 20, runs = 1000,scenario = scenarios[i], dispersal = dispersal[i],specRate = 2,density = density[i],environment = environment[i], fitnessBaseMortalityRatio = 5, densityCut = 1,seed = 1000)}

## Run the simulations
simulationOut <- runSimulationBatch(pars, parallel = "auto")

## Calculate null models
nullMeta <- calculatePhylogeneticDispersion(simulationOut, plotlengths = c(2,3,4,5,6,7,8,9,10), reduce = TRUE)

## Now we can define the positions and labels of the plots.
## Note that here the 28 simulations need to be covered in the right order.
## In this example there will be 7 rows and 4 collumns to represent the 28 cases.
positions <- list(x= c(1:7),
   y = c(4, 1:3), # this is beacause dispersal = "global" is equal to dispersal = "0"
   xname = c("neutral", "1", "0.75", "0.5", "0.25", "0", "both"),
   yname = c("global", "0.5", "2", "4"))

## Now you can plot the results.
plotPD(pvalues = nullMeta, positions = positions, title = "Null Meta")
}

