% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPhylogeneticDispersion.R
\name{plotPhylogeneticDispersion}
\alias{plotPhylogeneticDispersion}
\title{Plot Phylogenetix Dispersion}
\usage{
plotPhylogeneticDispersion(pvalues, positions = NULL, title = "P-values",
  multiple = T)
}
\arguments{
\item{pvalues}{List of p-values as generated by \code{\link{calculatePhylogeneticDispersion}}.}

\item{positions}{List, determines where the plots should be placed and how they should be labeled}

\item{title}{String, determines the title of the plot}
}
\value{
A plot
}
\description{
Plotting function for the results of \code{\link{calculatePhylogeneticDispersion}}. Usefull for plotting different scenarios.
}
\examples{
## Define a parameterset. Note that the setup here is duplicated in the
## names in the positions argument.
runs = 2000
dispOptions = 4
fitOptions = 7
 
dispersal =  rep(c(0,0.5, 2,4), each = fitOptions)
density = rep(c(0,seq(0,1,len = fitOptions-2),1),dispOptions)
environment = rep(c(0,seq(1,0,len = fitOptions-2),1),dispOptions)
modes <- ifelse(dispersal == 0, "global", "local")
scenarios <- paste(modes, " dens=", density, " env=", environment, sep = "")
pars <- list()
for (i in 1:length(scenarios)){pars[[i]] = createCompletePar(x = 20,y = 20, runs = runs,scenario = scenarios[i], dispersal = dispersal[i],specRate = 2,density = density[i],environment = environment[i], fitnessBaseMortalityRatio = 5, densityCut = 1,seed = 1000)}

## Run the simulations
simulationOut <- runSimulationBatch(pars, parallel = "auto")

## Calculate null models
nullMeta <- calculatePhylogeneticDispersion(simulationOut, plotlengths = c(2,3,4,5,6,7,8,9,10), reduce = TRUE)

## Define the positions where the plots should be drawn
positions <- list(x= c(1, seq(2.5,6.5,1), 8),
   y = c(4, 1:3),
   xname = c("neutral", "1", "0.75", "0.5", "0.25", "0", "both"),
   yname = c("global", "0.5", "2", "4" ))

## Plot the results
plotPhylogeneticDispersion(pvalues = nullMeta, positions = positions, title = "Null Meta", multiple = TRUE) 
}

