% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/communityMetrics.R
\name{rac}
\alias{rac}
\title{Rank-abundance relation}
\usage{
rac(matrix, plot = "line")
}
\arguments{
\item{matrix}{Integer matrix representing the species community}

\item{plot}{Logical determining whether to plot the RAC or not}
}
\value{
A dataframe containing the ranked abundances, sorted by ascending rank.
}
\description{
Calculates the rank-abundance pattern for a community matrix.
}
\details{
The higest abundance is assigned rank "1" while the lowest is assigned the rank corresponding with species richness.
}

