% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{RcppModel}
\alias{RcppModel}
\title{Model of neutral community assembly}
\usage{
RcppModel(dim, specRate, runs)
}
\arguments{
\item{dim}{The edge-length of the square metacommunity matrix}

\item{specRate}{The speciation rate in total speciation events per generation}

\item{runs}{The number of generations the model runs through}

\item{seed}{The random seed (do not change this option if you previously set the seed!)}
}
\description{
Runs a simple model of species community assembly under neutral conditions as defined by Hubbel.
}
\details{
Be careful with the dimensions you choose. Large grids and many generations may take very long to compute. You can assume to have reached the equilibrium state of the model after dim^2/2.
}

