% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/runSimulationBatch.R
\name{runSimulationBatch}
\alias{runSimulationBatch}
\title{Batch runner}
\usage{
runSimulationBatch(pars, parallel = F, backup = FALSE)
}
\arguments{
\item{pars}{A single set of parameters as created by \code{\link{createCompletePar}} or a list of parameter sets}

\item{parallel}{The number of cores to be used for parallel computing (default null is no parallelization)}

\item{backup}{Logical determining whether the results of the individual scenario runs should be saved as a workspace image (advised if the simulation takes a long time, or if the individual scenarios vary greatly in runtime. Default is FALSE)}
}
\value{
A list containing numerical matrices of species distribution, local trait values and the environment for each scenario (each cell in a matrix represents an individual of a species), plus a phylogeny object of class "phylo" for each scenario.
}
\description{
A model of species community assembly under different assembly mechanisms, using parallel computing to make use of multi core cpus and clusters in order to reduce computation time.
}
\details{
This function uses the \code{\link{foreach}} and \code{\link{doParallel}} package to compute the model scenarios parallel on several cores. If you want to keep working on your computer make sure to reserve at least one core for your other endevors (by assigning n-1 cores to the function). By default all cores are employed to ensure maximum speed.\cr \cr The phylogeny is passed to R in the newick format and parsed to an object of class "phylo" with the function \code{\link[ape]{read.tree}} from the \code{\link{ape}} package.
}

