% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/speciesArea.R
\name{sac}
\alias{sac}
\title{Species-area relation}
\usage{
sac(matrix, area = NULL, rep = 50, plot = T)
}
\arguments{
\item{matrix}{The species community}

\item{area}{a single value or a vector determining the size(s) of the subplots that for which areas are calculated. If not provided, 10 sizes logarithmically will be chosen}

\item{rep}{The number of repetitions per size to calculate the mean}

\item{plot}{Logical determining whether to plot the SAC or not}
}
\value{
A list containing the mean species richness for each size and the respective standard deviation
}
\description{
Calculates the species-area pattern for a community matrix.
}

