% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/setLeipzigParameters.R
\name{setLeipzigParametersXML}
\alias{setLeipzigParametersXML}
\title{Set Model parameters (Leipzig)}
\usage{
setLeipzigParametersXML(scenarios, x, y, runs, dispersal, nSpec, specRate,
  density, densityStrength, location)
}
\arguments{
\item{scenarios}{Character vector containing the names of the scenarios to be run}

\item{x}{Dimension of the model landscape in x-direction}

\item{y}{Dimension of the model landscape in y-direction}

\item{runs}{Number of generations the model runs over}

\item{dispersal}{1 = Gloabl dispersal, 2 = Nearest neighbor dispersal}

\item{nSpec}{Number of initial Species in the model landscape}

\item{specRate}{Numerical the number of species introduced into the community during each generation}

\item{density}{Logical determining whether or not density dependence influences the model}

\item{densityStrength}{Numerical (0.0 - 1.0) determining the strength of the density dependence}

\item{location}{A character vector specifing the folder to save the .xml file in (the filename is always parameters.xml and does not need to be set.)}
}
\value{
Creates a .xml file at the specified location
}
\description{
Creates a .xml file of the desired model parameters to be loaded into the model.
}

