% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/speciesRichness.R
\name{specRich}
\alias{specRich}
\title{Species richness}
\usage{
specRich(simu, which.simulation = NULL)
}
\arguments{
\item{simu}{Object of class "Phylosim", usually consisting out of several lists. Needs at least the spatial distribution of the species stored in a matrix ($specMat)}

\item{which.simulation}{Integer, determines which simulation should be used. Only useful if your Phylosim object contains more than one result.}
}
\value{
An integer value for the species richness
}
\description{
Calculate the total species richness of a community
}

