% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/parCreator.R
\name{createCompletePar}
\alias{createCompletePar}
\title{Parameter Generator}
\usage{
createCompletePar(x = 50, y = 50, dispersal = "global", runs = 100,
  specRate = 1, density = F, environment = F,
  fitnessActsOn = "mortality", fitnessBaseMortalityRatio = 10,
  densityCut = 1, seed = NULL, type = "base", modes = NULL,
  scenario = NULL)
}
\arguments{
\item{x}{Integer, Dimension of the model landscape in x-direction}

\item{y}{Integer, Dimension of the model landscape in y-direction}

\item{dispersal}{Integer. Type 0 or "global" for global dispersion. For local dispersion all integers >=1 set the dispersal distance.}

\item{runs}{Integer, Number of generations the model runs over. You can also define a sequence of runs. Then the intermediate and end results are saved.}

\item{specRate}{Integer, Number of Individuals introduced to the community in each generation}

\item{density}{Logical, determining whether or not density dependence influences the model}

\item{environment}{Logical, determining whether or not the environment influences the model}

\item{fitnessActsOn}{Character, determining how the fitness influences the individuals. Possible inputs are "mortality" (default), "reproduction" or "both"}

\item{fitnessBaseMortalityRatio}{Integer, TODO}

\item{densityCut}{Integer, defines the effective range of the competition (ignored if density = FALSE)}

\item{seed}{numerical, sets the random seed}

\item{type}{Character, determining which model should be used. "base" is running the default model. Other possibilities are "Leipzig" and "Rneutral" which will run a neutral model purely in R.}
}
\value{
A List with parameters
}
\description{
Function to create a list with parameters for biogeographical simulations with \code{\link[PhylGeo]{runSimulation}} or \code{\link[PhylGeo]{runSimulationBatch}}
}
\details{
If type = "Rneutral" the model will run entirely in R. This model is to be seen only for test and teaching purpose. To be used in practice it is far too slow. Also the output is reduced in comparision with the other models. Only the species landscape and the parameter settings will be displayed in the output.
}

