% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rWrappers.R
\name{fullMod}
\alias{fullMod}
\title{Species community Simulation}
\usage{
fullMod(x, y, dispersal, runs, specRate, density, environment, neutral)
}
\arguments{
\item{x}{Dimension of the model landscape in x-direction}

\item{y}{Dimension of the model landscape in y-direction}

\item{dispersal}{1 = Gloabl dispersal, 2 = Nearest neighbor dispersal}

\item{runs}{Number of generations the model runs over}

\item{density}{Logical determining whether or not density dependence influences the model}

\item{environment}{Logical determining whether or not the environment influences the model}

\item{neutral}{Logical determining whether or not the model is run under neutral conditions (overrides density and environment)}
}
\value{
a numerical matrix (each cell represents an individual of a species)
}
\description{
A model of species community assembly under different assembly mechanisms
}

