% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/modelSingle.R
\name{fullMod}
\alias{fullMod}
\title{Species community Simulation}
\usage{
fullMod(x = NULL, y = NULL, dispersal = NULL, runs = NULL,
  specRate = NULL, density = NULL, environment = NULL,
  mortalityFitness = NULL, mortalityStrength = NULL,
  reproductiveFitness = NULL, neutral = NULL, dispersalCut = 2,
  densityCut = 1, seed = NULL, saveLocation = NULL)
}
\arguments{
\item{x}{Integer, Dimension of the model landscape in x-direction}

\item{y}{Integer, Dimension of the model landscape in y-direction}

\item{dispersal}{Integer, 1 = Gloabl dispersal, 2 = Nearest neighbor dispersal}

\item{runs}{Integer, Number of generations the model runs over}

\item{specRate}{Integer, Number of Individuals introduced to the community in each generation}

\item{density}{Logical, determining whether or not density dependence influences the model}

\item{environment}{Logical, determining whether or not the environment influences the model}

\item{mortalityFitness}{Logical, determining whether or not the individual fitness influences mortality}

\item{mortalityStrength}{Integer determining the influence of fitness on mortality (see Details)}

\item{reproductiveFitness}{Logical, determining whether or not the individual fitness influences reproduction}

\item{neutral}{Logical, determining whether or not the model is run under neutral conditions (overrides density and environment)}

\item{dispersalCut}{Integer, defines the dispersal distance for local dispersal (ignored if dispersal = 1 (Global))}

\item{densityCut}{Integer, defines the effective range of the competition (ignored if density = FALSE)}

\item{seed}{numerical, sets the random seed}

\item{saveLocation}{Path of the folder to save the backup files to}
}
\value{
A List containing sveral numerical matrices with species (each cell represents an individual of a species), environment and different traits as well as the phylogeny as an object of class 'phylo'.
}
\description{
A model of species community assembly under different assembly mechanisms
}
\details{
If mortalityFitness = TRUE, the number of reproduction events per generation is doubled. The mortality strength is an Integer value which determines the number of events per generation in which the mortality depends on fitness. \cr E.g. if mortalityStrengt = 100 for every 100 events with fitness dependent mortality there is 1 run with random mortality. \cr This way the random base mortality can be increased or decreased. The higher the frequency of the random base mortality the more neutral the conditions get.
}
\examples{
# Run the model
myModel <- fullMod(x = 50, y = 50, dispersal = 1, runs = 100, specRate = 1.0, density = FALSE, environment = FALSE, neutral = TRUE, seed = 1500)

 # Look at the phylogeny (requires package 'ape')
 require(ape)

 # Get the phylogeny
 phylogeny <- myModel$phylogeny

 # Only extant taxa
 extantPhylogeny <- drop.fossil(phylogeny)

 plot(extantPhylogeny)

 #Look at the species area relation

 species <- myModel$specMat
 sac(area = c(1,10,100,1000), matrix = species, rep = 100, plot= T)
}

