% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localPlots.R
\name{localPlots}
\alias{localPlots}
\title{Create Subplots}
\usage{
localPlots(simu, which.simulation = NULL, size, n, community = F)
}
\arguments{
\item{simu}{Simulation output of the class "Phylosim", usually consisting out of several lists. Needs to contain at least a species matrix ($specMat)}

\item{which.simulation}{defines which simulation run to choose in case you defined to save at multiple time steps. The default is the last one.}

\item{size}{A single value or a vector determining the size(s) of the generated subplots}

\item{n}{The number of subplots to be generated}

\item{community}{Logical, determining whther to generate a communiy table or not. default is FALSE}
}
\value{
A list of subplots and (if coomunity = T) a community table with plots in rows and species in collumns
}
\description{
Generates subplots from a metacommunity, simulating field conditions
}

