% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nullModel.R
\name{nullModel}
\alias{nullModel}
\title{Null Model}
\usage{
nullModel(simu, which.simulation = NULL, localPlotSize, numberOfPlots,
  repetitions)
}
\arguments{
\item{simu}{an object of class "phylosim" as created by \code{\link{runSimulation}} or \code{\link{runSimulationBatch}}}

\item{which.simulation}{Integer, determines which result should be used. This argument is only usefull if interim steps are saved in the Phylosim object. By default (NULL), the end result is used.}

\item{localPlotSize}{number of grid cells of the plots drawn within the metacommunity}

\item{numberOfPlots}{number of plots drawn within the metacommunity}

\item{repetitions}{number of generated null model plots}
}
\value{
A numeric vector with pValues for each plot in the observed metacommunity
}
\description{
Generates a null model to compare the observed results against.
}
\examples{
### Usage with runSimulation
## Define a parameter set
par <- createCompletePar(x = 50, y = 50, dispersal = 0 , runs = c(500,1000), density = 1, environment = 0.5, specRate = 1)

## Run Model
simu <- runSimulation(par)

## Compare to null model
nullModel(simu, localPlotSize = 100, numberOfPlots = 10, repetitions = 10)

## Usage with runSimulationBatch
## Define two (or multiple) parameter sets
par1 <- createCompletePar(x = 50, y = 50, dispersal = 0 , runs = c(500,1000),
        density = 0, environment = 0.5, specRate = 1)
par2 <- createCompletePar(x = 50, y = 50, dispersal = 0 , runs = c(500,1000),
        density = 1, environment = 0.5, specRate = 1)

## Merge the parameter sets. It is important to note, that the funktion
## needs a list of parameter sets.
par <- list(par1,par2)

## Run the model
simubatch <- runSimulationBatch(par, parallel=T)

## Compare results to null model
nullModel(simubatch[[1]],localPlotSize = 100, numberOfPlots = 10, repetitions = 10)
nullModel(simubatch[[2]],localPlotSize = 100, numberOfPlots = 10, repetitions = 10)
}

