% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconstructPhylogeny.R
\name{phyloReconstruct}
\alias{phyloReconstruct}
\title{Phylogeny Reconstruction}
\usage{
phyloReconstruct(simu, which.simulation = NULL, ...)
}
\arguments{
\item{simu}{A object of class Phylosim}

\item{which.simulation}{Integer, determines which simulation should be used. Only useful if your Phylosim object contains more than one result. By default the last result is chosen.}

\item{...}{Additional arguments to be passed to function \code{\link{hclust}}}
}
\value{
An object of class 'phylo'
}
\description{
reconstructs a phylogeny from a neutral trait matrix as returned by \code{\link{runSimulation}}
}
\examples{
 
## Run a simulation
par <- createCompletePar(x = 50, y = 50, dispersal = "global" , runs = 500,
        density = 1, environment = 0.5, specRate = 1)

simu <- runSimulation(par)

## Reconstruct phylogeny
rePhyl <- phyloReconstruct(simu)

## Compare to the real phylogeny (requires package 'ape')
par(mfrow=c(1,2))
plot(drop.fossil(simu$Output[[1]]$phylogeny)
plot(rePhyl)
}

