% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotSpatialPhylo.R
\name{plotSpatialPhylo}
\alias{plotSpatialPhylo}
\title{Plot of Phylogeny and species Landscape}
\usage{
plotSpatialPhylo(landscape, phylogeny, plot = "both")
}
\arguments{
\item{landscape}{the species landscape}

\item{phylogeny}{the corresponding (extant) phylogeny}

\item{plot}{defines what to plot "both" plots the landscape and phylogeny side-by-side}
}
\description{
Plots the phylogeny and species landscape, colored by taxon
}

