% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotSpatialPhylo.R
\name{plotSpatialPhylo}
\alias{plotSpatialPhylo}
\title{Plot of phylogeny and species landscape}
\usage{
plotSpatialPhylo(simu, plot = "both", plotTraits = T, col = "phylodist",
  main = "", time = NULL)
}
\arguments{
\item{simu}{simulation output, consisting of a list contain at least $specMat with the spatial species matrix, and $phylogeny with the phylogeny}

\item{plot}{defines what to plot. "both" plots the landscape and phylogeny side-by-side, other options are "landscape" and "phylogeny"}

\item{plotTraits}{defines if the traits should be plotted}

\item{col}{defines the color, options are "phylodist" (default) and "equidist"}

\item{main}{defines the title of the plot}

\item{time}{defines which simulation run to choose in case you defined to save at multiple time steps. The default is the last one.}

\item{phylogeny}{the corresponding (extant) phylogeny}
}
\description{
Plots the phylogeny and species landscape, colored by taxon
}

