% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/traitPlot.R
\name{plotTraitDistribution}
\alias{plotTraitDistribution}
\title{Trait plots}
\usage{
plotTraitDistribution(simu, type = "hist", time = NULL)
}
\arguments{
\item{landscape}{the species landscape}

\item{phylogeny}{the corresponding (extant) phylogeny}

\item{plot}{defines what to plot. "both" plots the landscape and phylogeny side-by-side}
}
\description{
Plot traits of simulation outputs
}

