% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotTraitDistribution.R
\name{plotTraitDistribution}
\alias{plotTraitDistribution}
\title{Trait plots}
\usage{
plotTraitDistribution(simu, type = "hist", which.simulation = NULL)
}
\arguments{
\item{simu}{Simulation output of the class "Phylosim", usually consisting out of several lists. Needs to contain at least the three trait matrices ($traitMat;$compMat;$neutMat) and the environment matrix ($envMat)}

\item{which.simulation}{defines which simulation run to choose in case you defined to save at multiple time steps. The default is the last one.}
}
\description{
Plots trait-histograms, trait ~ Environment relationship and the spatial distribution of the different traits
}

