% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotRAC.R
\name{rac}
\alias{rac}
\title{Rank Abundance Curve}
\usage{
rac(simu, which.simulation = NULL, plot = "line")
}
\arguments{
\item{simu}{Simulation output of the class "Phylosim", usually consisting out of several lists. Needs at least the spatial distribution of the species stored in a matrix ($)}

\item{plot}{determining whether to plot the RAC as "line"(default) or "bar".}
}
\value{
A dataframe containing the ranked abundances, sorted by ascending rank.
}
\description{
Plots the Rank Abundance Curve for a given community.
}
\details{
The highest abundance is assigned rank "1" while the lowest is assigned the rank corresponding with species richness.
}

