% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/runSimulation.R
\name{runSimulation}
\alias{runSimulation}
\title{Species community Simulation}
\usage{
runSimulation(par)
}
\description{
A model of species community assembly under different assembly mechanism
}
\examples{
# Run the model
myModel <- fullMod(x = 50, y = 50, dispersal = 1, runs = 100, specRate = 1.0, density = FALSE, environment = FALSE, neutral = TRUE, seed = 1500)

 # Look at the phylogeny (requires package 'ape')
 require(ape)

 # Get the phylogeny
 phylogeny <- myModel$phylogeny

 # Only extant taxa
 extantPhylogeny <- drop.fossil(phylogeny)

 plot(extantPhylogeny)

 #Look at the species area relation

 species <- myModel$specMat
 sac(area = c(1,10,100,1000), matrix = species, rep = 100, plot= T)
}

