% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/runSimulation.R
\name{runSimulation}
\alias{runSimulation}
\title{Species Community Simulation}
\usage{
runSimulation(par)
}
\arguments{
\item{par,}{a list of parameters created with \link{createCompletePar}}
}
\value{
An object of class "Phylosim". This objet contains the species matrix, the trait matrix, the environmental matrix, the competition matrix and the neutral matrix, as well as the phlogeny and the parameter settings of the simulation.
}
\description{
A model of species community assembly under different assembly mechanisms.
}
\details{
If your parameterset contains more than one runs argument, each interim step is saved in the Phylosim object. \cr\cr For larger simularions consider \link{runSimulationBatch} to make use of parallel computing.
}
\examples{
# Define a parameter set
par <- createCompletePar(x = 50, y = 50, dispersal = F , runs = c(500,1000),
        density = T, environment = 0.5, specRate = 1)

# Run the model
simu <- runSimulation(par)

# Look at the phylogeny (requires package 'ape')
require(ape)

# Get the phylogeny of the last run. In this example this is after 1000 runs.
phylogeny <- simu[[2]]$phylogeny

# Only extant taxa
extantPhylogeny <- drop.fossil(phylogeny)

# Display the results
plot(extantPhylogeny)

#Look at the species area relation
sac(area = c(1,10,100,1000), matrix = simu[[2]]$specMat, rep = 100, plot= T)
}

