% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotSAC.R
\name{sac}
\alias{sac}
\title{Species Area Curve}
\usage{
sac(simu, which.simulation = NULL, area = NULL, rep = 50, plot = T)
}
\arguments{
\item{simu}{Simulation output of the class "Phylosim", usually consisting out of several lists. Needs at least the spatial distribution of the species stored in a matrix ($specMat)}

\item{which.simulation}{Integer, determines which simulation should be used. Only useful if your Phylosim object contains more than one result.}

\item{area}{a single value or a vector determining the size(s) of the subplots that for which areas are calculated. If not provided, 10 sizes logarithmically will be chosen}

\item{rep}{The number of repetitions per size to calculate the mean}

\item{plot}{Logical determining whether to plot the SAC or not}
}
\value{
A list containing the mean species richness for each size and the respective standard deviation
}
\description{
Plots the species area curve for a given community.
}
\examples{
#Plot SAC curve for neutral model and global dispersion
simu <- simu.neutral.global
sac(simu = simu)

#Plot SAC curve for neutral model and local dispersion
simu <- simu.neutral.local
sac(simu = simu)

#Plot SAC curve for environment and competition model and local dispersion
simu <- simu.envcom.local
sac(simu = simu)

#Plot SAC curve with random plotsize
simu <- simu.envcom.local
sac(simu=simu, area = sort(sample(c(10:1000), size = 10)))

#Plot SAC curve with different repititions
simu <- simu.envcom.local
par(mfrow=c(3,1))
sac(simu=simu, rep = 3)
sac(simu=simu, rep = 30)
sac(simu=simu, rep = 30)
}

