% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/speciesArea.R
\name{sac}
\alias{sac}
\title{Species-area relation}
\usage{
sac(area, matrix, rep = 5, plot = T)
}
\arguments{
\item{area}{A single value or a vector determining the size(s) of the generated subplots}

\item{matrix}{The species community}

\item{rep}{The number of repetitions per size to calculate the mean}

\item{plot}{Logical determining whether to plot the SAC or not}
}
\value{
A list containing the mean species richness for each size and the respective standard deviation
}
\description{
Calculates the species-area pattern for a community matrix.
}

