% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rFunctions.R
\name{setModelParametersXML}
\alias{setModelParametersXML}
\title{Set Model parameters}
\usage{
setModelParametersXML(scenarios, x, y, runs, dispersal, specRate, density,
  environment, neutral, location, dispersalCut, densityCut, seed)
}
\arguments{
\item{scenarios}{Character vector containing the names of the scenarios to be run}

\item{x}{Dimension of the model landscape in x-direction}

\item{y}{Dimension of the model landscape in y-direction}

\item{runs}{Number of generations the model runs over}

\item{dispersal}{1 = Gloabl dispersal, 3 = Local dispersal}

\item{specRate}{Numerical the number of species introduced into the community during each generation}

\item{density}{Logical determining whether or not density dependence influences the model}

\item{environment}{Logical determining whether or not the environment influences the model}

\item{neutral}{Logical determining whether or not the model is run under neutral conditions (overrides density and environment)}

\item{location}{A character vector specifing the folder to save the .xml file in (the filename is always parameters.xml and does not need to be set.)}

\item{dispersalCut}{Integer defining the maximum dispersal distance of the kernel}

\item{densityCut}{Integer defining the range for the density dependence}

\item{seed}{Integer setting the random seed for the model}
}
\value{
Creates a .xml file at the specified location
}
\description{
Creates a .xml file of the desired model parameters to be loaded into the model.
}

