% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotRAC.R
\name{specRich}
\alias{specRich}
\title{Species richness}
\usage{
specRich(simu, which.simulation = NULL)
}
\arguments{
\item{matrix}{A square matrix containing a species community (one individual per grid cell)}
}
\value{
An integer value for the species richness
}
\description{
Calculate the total species richness of a community
}

